Update to latest dev

+++ scripts/download.sh
@@ -5,7 +5,16 @@
 
 set -eo pipefail
 
-cd openwrt
+if [ -z "${OPENWRT_COMPILE_DIR}" ] || [ -z "${OPENWRT_CUR_DIR}" ] || [ -z "${OPENWRT_SOURCE_DIR}" ]; then
+  echo "::error::'OPENWRT_COMPILE_DIR', 'OPENWRT_CUR_DIR' or 'OPENWRT_SOURCE_DIR' is empty" >&2
+  exit 1
+fi
+
+[ "x${TEST}" != "x1" ] || exit 0
+
+cd "${OPENWRT_CUR_DIR}"
+make defconfig
+make oldconfig
 make download -j8
 find dl -size -1024c -exec ls -l {} \;
 find dl -size -1024c -exec rm -f {} \;
